refactor: separate responsibility for files
