+++ src/chattypes.py
@@ -1,0 +1,6 @@
+from typing import TypedDict
+
+
+class Conversation(TypedDict):
+    role: str
+    content: str
